Add `Option` and unit variant support valence_nbt serde
